Moving specs arounds (Shoryuken => Options)
